+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 medialibrary CHANGELOG
 ======================
 
+## CURRENT
+ - [medialibrary-140](https://github.com/cjcodeproj/medialibrary/issues/140) Output a catalog title
+
 ## RELEASE 0.2.1
  - [medialibrary-125](https://github.com/cjcodeproj/medialibrary/issues/125) Create simple example documentation
  - [medialibrary-127](https://github.com/cjcodeproj/medialibrary/issues/127) Consider deprecating random_sample_list()